add adam optimizer, extract APIs to separate files
